Finished landing page for now. Will add an additional section and possibly some new features to the footer in the future. Also renamed some of my components.
