+++ src/components/HomeSectionArticles.jsx
@@ -5,7 +5,7 @@ import HomeArticle from './HomeArticle';
 class SectionArticles extends React.Component {
     render() {
         return (
-            <section className="main-section-wrapper">
+            <section className="main-section-wrapper" id="most-viewed-articles">
                 <h2 className="section-title">Most Viewed Articles</h2>
                 <hr className=" most-viewed-hr"/>
                 <ul className="flex-container home-article-container" id="home-article-container">
