+++ src/sass/base.scss
@@ -17,10 +17,16 @@
 
 @media (max-width: 576px) {
     #home-article-container {
-        flex-direction: column;
-        justify-content: center;
-        justify-items: center;
+        flex-flow: column wrap;
         padding-left: 53px;
+
+    }
+}
+
+@media (min-width: 577px) {
+    #home-article-container {
+        justify-content: space-evenly;
+        max-width: 350px;
     }
 }
 
@@ -103,6 +109,7 @@ body {
 
                 .page-title {
                     font-weight: 700;
+                    letter-spacing: 1px;
                     margin: 10 auto 0 auto;
                     text-align: center;
                     color: color(primary-blue);
@@ -147,11 +154,17 @@ body {
 
                 // margin: 30px auto;
 
+                #most-viewed-articles {
+                    margin-top: 0;
+                }
+
                 .main-section-wrapper {
                     margin: 30px auto;
 
                     .section-title {
                         text-align: center;
+                        color: color(primary-blue);
+                        letter-spacing: 1px;
                     }
 
                     .most-viewed-hr {
@@ -163,12 +176,21 @@ body {
                         list-style: none;
                     }
 
+                    #home-article-container {
+                        justify-content: space-evenly;
+                        justify-items: center;
+                        margin: 25px 0 0 0;
+                        letter-spacing: 1.1px;
+                        font-size: 14px;
+                    }
+
                     .home-article-container {
+                        // list-style: none;
                         flex-direction: column;
+                        margin: 50px 0;
 
                         .article-link {
                             text-decoration: none;
-                            border-bottom: 1px solid #000;
                         }
 
                     }
@@ -199,6 +221,7 @@ body {
     
                                 h2 {
                                     color: color(primary-yellow);
+                                    letter-spacing: 1px;
                                 }
     
                                 h3 {
